Fixed an issue regarding mouse wheel zoom. Added prerelease brand. Updated to version 1.0.8.0

+++ FolderSizeWin32/FolderSizeWin32/MainWindow/main_window.cpp
@@ -1072,7 +1072,7 @@ namespace main_window
             break;
          case WM_MOUSEWHEEL:
             {
-               auto scroll = static_cast<short> (HIWORD (w_param)) / WHEEL_DELTA;
+               auto scroll = static_cast<double> (static_cast<short> (HIWORD (w_param))) / WHEEL_DELTA;
                auto mouse_coord = w::get_client_mouse_coordinate (hwnd, l_param);
 
                if (s_state.get () && w::is_inside (folder_tree_rect, mouse_coord))
